phase0_streamlit/requirements.txt
streamlit>=1.30.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.0
